requirements.txt
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
Deprecated==1.2.14
Flask==2.0.1
gcloud==0.18.3
googleapis-common-protos==1.60.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauth2client==4.1.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-http-client==3.3.7
python-jwt==4.0.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==2.2.0
urllib3==1.26.17
Werkzeug==2.2.2
wrapt==1.15.0
